help

+++ test/01-HeyTriangle/genShape.h
@@ -4,5 +4,5 @@ struct Geometry;
 
 Geometry makeNGon(size_t N, float r);
 
-Geometry makeCheckerboard(int dim, int size);
+Geometry makeCheckerboard(int dim, float size);
